requirements.txt
fastapi==0.100.0
uvicorn==0.23.2
openai==1.3.5
python-dotenv==1.0.0